+++ test/insert.test.ts
@@ -1,6 +1,13 @@
 import { query } from '../src'
 
-import { client, events, users } from './helpers'
+import {
+  client,
+  events,
+  users,
+  eventTypes,
+  eventTypesWithEnum,
+  EventTypeEnum,
+} from './helpers'
 
 describe('insert', () => {
   async function queryUsers(ids: number[]) {
@@ -207,6 +214,70 @@ describe('insert', () => {
         eventType: 'A',
       } as any),
     ).rejects.toThrow('expected a data:string attribute')
+  })
+
+  test('string literal union data', async () => {
+    const res = await query(eventTypes).insertOne(client, {
+      active: false,
+      description: 'test',
+      type: 'D',
+    })
+
+    expect(res).toEqual({
+      active: false,
+      description: 'test',
+      type: 'D',
+    })
+  })
+
+  test('string literal union data validation', async () => {
+    await expect(
+      query(eventTypes).insertOne(client, {
+        active: false,
+        description: 'test',
+        type: 'U',
+      } as any),
+    ).rejects.toThrow(
+      "column type - expected a string of A,B,C,D,E,X but got: 'U'",
+    )
+  })
+
+  test('enum data', async () => {
+    const res = await query(eventTypesWithEnum).insertOne(client, {
+      active: false,
+      description: 'test',
+      type: EventTypeEnum.TypeD,
+    })
+
+    expect(res).toEqual({
+      active: false,
+      description: 'test',
+      type: 'D',
+    })
+  })
+
+  test.only('enum data validation', async () => {
+    await expect(
+      query(eventTypesWithEnum).insertOne(client, {
+        active: false,
+        description: 'test',
+        type: 'TypeA',
+      } as any),
+    ).rejects.toThrow(
+      "column type - expected a member of the enum { '0': 'TypeNumber', TypeA: 'A', TypeB: 'B', TypeC: 'C', TypeD: 'D', TypeE: 'E', TypeX: 'X', TypeNumber: 0 } but got: 'TypeA'",
+    )
+  })
+
+  test.only('enum data validation - checking typescript reverse number enum mappings', async () => {
+    await expect(
+      query(eventTypesWithEnum).insertOne(client, {
+        active: false,
+        description: 'test',
+        type: 'TypeNumber',
+      } as any),
+    ).rejects.toThrow(
+      "column type - expected a member of the enum { '0': 'TypeNumber', TypeA: 'A', TypeB: 'B', TypeC: 'C', TypeD: 'D', TypeE: 'E', TypeX: 'X', TypeNumber: 0 } but got: 'TypeNumber'",
+    )
   })
 
   // TODO what about non-base tables?